hud doesn't working

+++ src/main/kotlin/ru/rikgela/russianmagic/client/gui/GuiManaHUD.kt
@@ -25,21 +25,26 @@ class GuiManaHUD(
         val y = minecraft.mainWindow.scaledHeight * yPos / 100
         var color = 0x337CFF
         if (mana.currentMana > mana.maxMana) {
-            val lambda = min((mana.currentMana - mana.maxMana).toFloat() / (mana.maxMana.toFloat() * 2F), 1F)
+            val lambda = min((mana.currentMana - mana.maxMana) / (mana.maxMana * 2F), 1F)
             val colorRed = (0x33 - ((0x33 - 0xFF)).toFloat() * lambda).toInt()
             val colorGreen = (0x7C - ((0x7C - 0x00)).toFloat() * lambda).toInt()
             val colorBlue = (0xFF - ((0xFF - 0x00)).toFloat() * lambda).toInt()
             color = colorRed * 0x10000 + colorGreen * 0x100 + colorBlue
         }
         fill(x, y, width + x, y - height, -0x1000000 or 0xC9CAB9)
-        fillGradient(x, y, (width * min(mana.currentMana.toFloat() / mana.maxMana.toFloat(), 1F)).toInt() + x, y - height, -0x1000000 or color, Color(-0x1000000 or 0x1145A1).darker().rgb)
+        fillGradient(x, y,
+                (width * min(mana.currentMana.toFloat() / mana.maxMana, 1F)).toInt() + x,
+                y - height,
+                -0x1000000 or color,
+                Color(-0x1000000 or 0x1145A1).darker().rgb)
     }
 
     fun printHud() {
         val x = minecraft.mainWindow.scaledWidth * xPos / 100
         val y = minecraft.mainWindow.scaledHeight * yPos / 100
         val mana = PlayerMana.fromPlayer(minecraft.player as PlayerEntity)
-        Minecraft.getInstance().fontRenderer.drawString(mana.currentMana.toString(), x.toFloat() + 3, y.toFloat() - 8, TextFormatting.WHITE.color!!)
+        Minecraft.getInstance().fontRenderer.drawString(
+                mana.currentMana.toString(), x.toFloat() + 3, y.toFloat() - 8, TextFormatting.WHITE.color!!)
     }
 
     companion object {
